Added Modules-README.yml
File in YML format to help keep track of the state of various modules. This first version was created mostly so others have a template of what information will be found in this file. Only a few of the modules (AlphaVantage, IEXCloud, HU, and YahooJSON) have current information represented.
Format and information subject to change as its use matures.
Added using a direct commit rather than branch, pull request, merge.