ASX website revamp, use JSON sources instead of HTML scraping.
October 2020 the ASX revamped their website with dynamic content for quotes
which prevented the previous HTML screen scraping from working, but exposed
a number of JSON data sources, two of which are used here.
The primary source returns data elements for almost all securities, but
does not return prices for certain security types (some bonds and exchange
traded products, options, and warrants), and returns an error for indices.
The alternate source returns less data elements, but provides usable quote
data for all the known exceptions, including indices.
This version will always call the primary source, and call the alternate if
a price is not returned by the primary.

Smart::Comments implemented to conform with the Hackers Guide:
https://github.com/finance-quote/finance-quote/blob/master/Documentation/Hackers-Guide