Add TRT suffix to AlphaVantage
When using online quotes in GnuCash, Toronto Stock Exchange prices for stocks like
MRT-UN.TO, HOT-UN.TO, CUF-UN.TO, PLZ-UN.TO, BPF-UN.TO are not being
populated.
The reason this happens is because in AlhaVantage these stocks are represented with
the TRT suffix (e.g. MRT-UN.TRT instead of MRT-UN.TO).
This pull request adds the TRT suffix and associates CAD to it.