Fix for updated Bloomberg website
This should fix #324

The module must now use cookies to retrieve the quote data successfully.
This is done using HTTP::CookieJar::LWP, which is a new (but small)
dependency.

The module now provides the 'name' label in addition to the existing
support for 'last', 'currency', 'symbol' and 'isodate'.

Conversion is also done from British pence to pounds where necessary.
The associated test now includes a lookup that uses this feature.