Fix for updated Bloomberg website
This should fix #360

At first it just looked like some of the HTML structure had changed,
requiring slightly different tag searches. However, at least for US
stocks, the price data now seems to be filled in using a script, so
the HTML as seen by F::Q is not the same as what can be saved from
a browser and the price may be missing.

Fortunately, the data is available in a nice JSON structure that is
embedded in a <script> tag, so it's possible to retrieve all of the
required data by extracting and parsing that JSON blob.