Add UKX:IND as a test symbol for Bloomberg.pm
This has a value large enough to be displayed on the website with
a comma (e.g. price of '6000' would be displayed as '6,000'). The
comma needs to be removed by the module so that the returned price
is a proper numeric value.

UKD:IND is the Financial Times Stock Exchange 100 Index, a stock
market index of 100 of the most highly capitalised blue chip
companies listed on the London Stock Exchange.