Fixed TesouroDireto.pm (again) using Tesouro Transparente
A few days after the last change to Tesouro Direto's, using
the csv from the redesigned site, they began showing CloudFlare's
javascript challenge on it, effectively breaking the module.

This commit changes the dataset used for quotes to Tesouro
Transparente's csv with historical quote data.

This dataset has a few downsides, being a bigger download
(full historical data since 2004) and having 1 day delay for
updates, but at least it's suposed to be used for machine
processing and, as such, hopefully we shouldn't have problems
with anti-bot measures.